base/nebmods: Replace local file-copy hack with my_fdcopy()
my_f{,d}copy() exists so we can copy files safely and have all
the error checking (and logging) done in one place. Implementing
a local version of it with less error-checking and logging is a
Bad Idea(tm).

While we're at it, we also make sure to release the output_file
variable if we fail to copy the file, plugging a small and largely
insignificant leak.

Signed-off-by: Andreas Ericsson <ae@op5.se>


Author: Andreas Ericsson <ae@op5.se>