base/nebmods: Plug a (tiny) leak
When loading eventbroker modules we have lots of error paths,
so there are lots of places where we'd want to free() the
constructed filename from which we loaded the module earlier.

This patch converts the buffer containing the constructed
filename from heap-allocated to stack-allocated. Doing so plugs
the one-time leak and, more importantly, makes valgrind shut
the fsck up.

Signed-off-by: Andreas Ericsson <ae@op5.se>